+++ src/graphql/root-schema.js
@@ -21,6 +21,7 @@ const rootTypeDefs = gql`
 	type RootQuery {
 		user(email: String): User
 		device(mac: String, secret: String): Device
+		devices(secrets: [String!]!): [Device!]!
 		ledStrip(mac: String!, ledStripId: String!): LedStrip!
 			@requiresAuth(acceptsOnly: [SELF, OWNER, MANAGER])
 		ledStrips(mac: String!): [LedStrip!]!